Removed one Email Screen
