+++ src/config/labels.js
@@ -3,12 +3,7 @@ const labels={
         MORE:'More',
         LOGIN_OR_SIGNUP:'Login / Signup'
     },
-    LOGINWITHOTP:{
-        SEND_OTP:'Send otp',
-        DIVIDER_OR  :'-------------------------  OR  -------------------------',
-        LOGIN_THROUGH_EMAIL:'Login with Email'
-    },
-    LOGINCONTAINER:{
+    LOGINSIGNUP:{
         NEVER_SHARE_YOUR_DETAILS:'We Will never share your email or phone number with anyone else.',
         DONT_HAVE_ACCNT:'Dont have an account ?',
         CREATE_NEW_ACCOUNT:'Create now',
@@ -16,13 +11,12 @@ const labels={
         AFTER_FILLED_ALL_DETAILS:'Submit',
         CLOSE_LOGINSIGNUP_PAGE:'Close',
     },
-    VAL_OTP:{
-        ENTER_PHONE_NUMBER:'Please Enter your Phone Number',
-        PHONE_NUMBER_CONTAINS:'Phone number contains only Numbers',
-        PHONE_NUMBERS_MUST_HAVE:'PhoneNumbers Must be 10 digits',
-        PHONE_NUMBERS_NOT_MORETHAN:'PhoneNumbers not more than 10 digits'
+    LOGINWITHOTP:{
+        SEND_OTP:'Send otp',
+        DIVIDER_OR  :'-------------------------  OR  -------------------------',
+        LOGIN_THROUGH_EMAIL:'Login with Email'
     },
-    VAL_CONTAINER:{
+    VAL_EMAIL:{
         ENTER_EMAIL_ID:'Please Enter Your Email ID',
         INVALID_EMAIL:'Invalid Email id  !',
         ENTER_EMAIL_IN_CORRECT_FORMAT:'(Please Enter Correct Format)',
@@ -30,6 +24,12 @@ const labels={
         PASSWORD_MUST_BE:'Password must be atleast 8 characters',
         PASSWORD_INCLUDES:'Password includes => ',
         PASSWORD_CONTAINS:'(Uppercase [A-Z] & lowercase [a-z] & Digits 0-9)'
-    }
+    },
+    VAL_OTP:{
+        ENTER_PHONE_NUMBER:'Please Enter your Phone Number',
+        PHONE_NUMBER_CONTAINS:'Phone number contains only Numbers',
+        PHONE_NUMBERS_MUST_HAVE:'PhoneNumbers Must be 10 digits',
+        PHONE_NUMBERS_NOT_MORETHAN:'PhoneNumbers not more than 10 digits'
+    },
 }
 export default labels
